# public/tools/schema/openapi_learning.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "🦎 GAD ",
    "description": "This is a sample REST server form jaktestowac.pl team.</br></br>Return to <a target='_blank' href='../tools/index.html' rel='noopener noreferrer' class='link'><strong>TOOLS🔨</strong></a>",
    "termsOfService": "http://jaktestowac.pl/",
    "version": "1.3.3"
  },
  "servers": [
    {
      "url": "/api"
    },
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "description": "Represents a user in the learning platform",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the user",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "User's login name",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "description": "User's email address",
            "example": "john.doe@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name",
            "example": "Doe"
          },
          "avatar": {
            "type": "string",
            "description": "URL to user's profile picture",
            "example": "https://example.com/avatars/johndoe.jpg"
          },
          "joinDate": {
            "type": "string",
            "format": "date",
            "description": "Date when user joined the platform",
            "example": "2023-01-15"
          },
          "role": {
            "type": "string",
            "enum": ["student", "instructor"],
            "description": "User's role in the platform",
            "example": "student"
          }
        }
      },
...
      "UserStats": {
        "type": "object",
        "properties": {
          "coursesInProgress": { "type": "integer" },
          "completedCourses": { "type": "integer" },
          "totalHours": { "type": "number" },
          "totalCertificates": { "type": "integer" },
          "lastUpdated": { "type": "string" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "NotFoundResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "InstructorStats": {
        "type": "object",
        "properties": {
          "totalCourses": { "type": "integer" },
          "totalStudents": { "type": "integer" },
          "averageRating": { "type": "number" },
          "totalRevenue": { "type": "number" }
        }
      },
...
    },
    "parameters": {
      "userIdParam": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "example": 1
        },
        "description": "User identifier"
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "User not authenticated",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "User not authorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NotFoundResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/learning/users/{userId}/stats": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user statistics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/userIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/learning/users/{userId}": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "tags": ["Users"],
        "summary": "Get user by ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/userIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/learning/instructor/stats": {
      "get": {
        "tags": ["Instructor"],
        "summary": "Get instructor statistics",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Instructor statistics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InstructorStats" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "403": { "$ref": "#/components/responses/ForbiddenError" }
        }
      }
    },
...
    "/learning/public/instructor/{instructorId}/stats": {
      "get": {
        "tags": ["Instructor"],
        "summary": "Get instructor's public statistics",
        "parameters": [
          {
            "name": "instructorId",
            "in": "path",
            "description": "Instructor ID",
            "required": true,
            "schema": { "type": "integer", "example": 2 }
          }
        ],
        "responses": {
          "200": {
            "description": "Instructor's public statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalCourses": { "type": "integer" },
                    "totalStudents": { "type": "integer" },
                    "averageRating": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/learning/public/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get public user profile",
        "parameters": [
          {
            "$ref": "#/components/parameters/userIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Public user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "avatar": { "type": "string" },
                    "joinDate": { "type": "string" },
                    "role": { "type": "string" },
                    "enrollments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "courseId": { "type": "integer" },
                          "courseTitle": { "type": "string" },
                          "progress": { "type": "integer" }
                        }
                      }
                    },
                    "certificates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "courseId": { "type": "integer" },
                          "courseTitle": { "type": "string" },
                          "issueDate": { "type": "string" }
                        }
                      }
                    },
                    "ratings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "courseId": { "type": "integer" },
                          "courseTitle": { "type": "string" },
                          "rating": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  }
}